+++ libraries/ghc-internal/src/GHC/Internal/Bignum/Integer.hs
@@ -133,6 +133,12 @@ module GHC.Internal.Bignum.Integer
     , integerBit
     , integerTestBit#
     , integerTestBit
+    , integerSetBit#
+    , integerSetBit
+    , integerClearBit#
+    , integerClearBit
+    , integerComplementBit#
+    , integerComplementBit
     , integerShiftR#
     , integerShiftR
     , integerShiftL#
@@ -706,6 +712,113 @@ integerTestBit# (IN x) i
 -- Fake 2's complement for negative values (might be slow)
 integerTestBit :: Integer -> Word -> Bool
 integerTestBit !i (W# n) = isTrue# (integerTestBit# i n)
+
+{- Note [INLINE for constant folding of bit operations]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+While there are no dedicated constant-folding rules for
+integerSetBit#/integerClearBit#/integerComplementBit#, we do INLINE them (and
+their Word-argument wrappers and the corresponding Bits Integer methods) to make
+the underlying primops accessible for constant folding, e.g. so that
+`clearBit (bit 0) 0 :: Integer` folds to `IS 0#`. Test T8832 checks the folding
+for all three operations.
+-}
+
+-- | Set the /n/-th bit.
+--
+-- Fake 2's complement for negative values (might be slow)
+--
+-- @since 10.201.0
+integerSetBit# :: Integer -> Word# -> Integer
+{-# INLINE integerSetBit# #-} -- See Note [INLINE for constant folding of bit operations]
+integerSetBit# n@(IS x) i
+   | isTrue# (i `ltWord#` (WORD_SIZE_IN_BITS## `minusWord#` 1##))
+   = IS (x `orI#` uncheckedIShiftL# 1# (word2Int# i))
+   | isTrue# (x >=# 0#)
+   = IP (bigNatSetBit# (bigNatFromWord# (int2Word# x)) i)
+   | True
+   = n
+integerSetBit# (IP x) i = IP (bigNatSetBit# x i)
+integerSetBit# (IN x) i = integerFromBigNatNeg#
+                             (bigNatAddWord#
+                                (bigNatClearBit# (bigNatSubWordUnsafe# x 1##) i)
+                                1##)
+
+-- | Set the /n/-th bit.
+--
+-- Fake 2's complement for negative values (might be slow)
+--
+-- @since 10.201.0
+integerSetBit :: Integer -> Word -> Integer
+{-# INLINE integerSetBit #-} -- See Note [INLINE for constant folding of bit operations]
+integerSetBit !i (W# n) = integerSetBit# i n
+
+-- | Clear the /n/-th bit.
+--
+-- Fake 2's complement for negative values (might be slow)
+--
+-- @since 10.201.0
+integerClearBit# :: Integer -> Word# -> Integer
+{-# INLINE integerClearBit# #-} -- See Note [INLINE for constant folding of bit operations]
+integerClearBit# n@(IS x) i
+   | isTrue# (i `ltWord#` (WORD_SIZE_IN_BITS## `minusWord#` 1##))
+   = IS (x `andI#` notI# (uncheckedIShiftL# 1# (word2Int# i)))
+   | isTrue# (x >=# 0#)
+   = n
+   | True
+   = IN (bigNatAddWord#
+           (bigNatSetBit#
+             (bigNatFromWord#
+                (minusWord# (int2Word# (negateInt# x)) 1##))
+             i)
+           1##)
+integerClearBit# (IP x) i = integerFromBigNat# (bigNatClearBit# x i)
+integerClearBit# (IN x) i = IN (bigNatAddWord#
+                                  (bigNatSetBit# (bigNatSubWordUnsafe# x 1##) i)
+                                  1##)
+
+-- | Clear the /n/-th bit.
+--
+-- Fake 2's complement for negative values (might be slow)
+--
+-- @since 10.201.0
+integerClearBit :: Integer -> Word -> Integer
+{-# INLINE integerClearBit #-} -- See Note [INLINE for constant folding of bit operations]
+integerClearBit !i (W# n) = integerClearBit# i n
+
+-- | Reverse the /n/-th bit.
+--
+-- Fake 2's complement for negative values (might be slow)
+--
+-- @since 10.201.0
+integerComplementBit# :: Integer -> Word# -> Integer
+{-# INLINE integerComplementBit# #-} -- See Note [INLINE for constant folding of bit operations]
+integerComplementBit# (IS x) i
+   | isTrue# (i `ltWord#` (WORD_SIZE_IN_BITS## `minusWord#` 1##))
+   = IS (x `xorI#` uncheckedIShiftL# 1# (word2Int# i))
+   | isTrue# (x >=# 0#)
+   = IP (bigNatSetBit# (bigNatFromWord# (int2Word# x)) i)
+   | True
+   = IN (bigNatAddWord#
+           (bigNatSetBit#
+              (bigNatFromWord# (minusWord# (int2Word# (negateInt# x)) 1##))
+              i)
+           1##)
+integerComplementBit# (IP x) i = integerFromBigNat# (bigNatComplementBit# x i)
+integerComplementBit# (IN x) i = integerFromBigNatNeg#
+                                    (bigNatAddWord#
+                                       (bigNatComplementBit#
+                                          (bigNatSubWordUnsafe# x 1##)
+                                          i)
+                                       1##)
+
+-- | Reverse the /n/-th bit.
+--
+-- Fake 2's complement for negative values (might be slow)
+--
+-- @since 10.201.0
+integerComplementBit :: Integer -> Word -> Integer
+{-# INLINE integerComplementBit #-} -- See Note [INLINE for constant folding of bit operations]
+integerComplementBit !i (W# n) = integerComplementBit# i n
 
 -- | Shift-right operation
 --